ddl: migrate test-infra to testify for ddl.testDBSuite5 (#33157)
close pingcap/tidb#32965
