+++ ddl/db_legacy_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/pingcap/tidb/ddl"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/errno"
-	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/meta/autoid"
@@ -47,14 +46,11 @@ import (
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/tablecodec"
-	ntestkit "github.com/pingcap/tidb/testkit"
-	"github.com/pingcap/tidb/util/admin"
 	"github.com/pingcap/tidb/util/dbterror"
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/testleak"
-	"github.com/pingcap/tidb/util/testutil"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/testutils"
 )
@@ -75,7 +71,6 @@ const (
 	waitForCleanDataInterval = time.Millisecond * 100
 )
 
-var _ = Suite(&testDBSuite5{&testDBSuite{}})
 var _ = SerialSuites(&testSerialDBSuite{&testDBSuite{}})
 
 const defaultBatchSize = 1024
@@ -139,35 +134,7 @@ func (s *testDBSuite) TearDownSuite(c *C) {
 	tearDownSuite(s, c)
 }
 
-type testDBSuite5 struct{ *testDBSuite }
 type testSerialDBSuite struct{ *testDBSuite }
-
-func (s *testDBSuite5) TestAddIndexWithDupIndex(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-
-	err1 := dbterror.ErrDupKeyName.GenWithStack("index already exist %s", "idx")
-	err2 := dbterror.ErrDupKeyName.GenWithStack("index already exist %s; "+
-		"a background job is trying to add the same index, "+
-		"please check by `ADMIN SHOW DDL JOBS`", "idx")
-
-	// When there is already an duplicate index, show error message.
-	tk.MustExec("create table test_add_index_with_dup (a int, key idx (a))")
-	_, err := tk.Exec("alter table test_add_index_with_dup add index idx (a)")
-	c.Check(errors.Cause(err1).(*terror.Error).Equal(err), Equals, true)
-	c.Assert(errors.Cause(err1).Error() == err.Error(), IsTrue)
-
-	// When there is another session adding duplicate index with state other than
-	// StatePublic, show explicit error message.
-	t := s.testGetTable(c, "test_add_index_with_dup")
-	indexInfo := t.Meta().FindIndexByName("idx")
-	indexInfo.State = model.StateNone
-	_, err = tk.Exec("alter table test_add_index_with_dup add index idx (a)")
-	c.Check(errors.Cause(err2).(*terror.Error).Equal(err), Equals, true)
-	c.Assert(errors.Cause(err2).Error() == err.Error(), IsTrue)
-
-	tk.MustExec("drop table test_add_index_with_dup")
-}
 
 func testGetTableByName(c *C, ctx sessionctx.Context, db, table string) table.Table {
 	dom := domain.GetDomain(ctx)
@@ -177,16 +144,6 @@ func testGetTableByName(c *C, ctx sessionctx.Context, db, table string) table.Ta
 	tbl, err := dom.InfoSchema().TableByName(model.NewCIStr(db), model.NewCIStr(table))
 	c.Assert(err, IsNil)
 	return tbl
-}
-
-func testGetSchemaByName(c *C, ctx sessionctx.Context, db string) *model.DBInfo {
-	dom := domain.GetDomain(ctx)
-	// Make sure the table schema is the new schema.
-	err := dom.Reload()
-	c.Assert(err, IsNil)
-	dbInfo, ok := dom.InfoSchema().SchemaByName(model.NewCIStr(db))
-	c.Assert(ok, IsTrue)
-	return dbInfo
 }
 
 func (s *testDBSuite) testGetTable(c *C, name string) table.Table {
@@ -363,234 +320,6 @@ func (s *testSerialDBSuite) TestDropTableOnTiKVDiskFull(c *C) {
 	tk.MustExec("drop table test_disk_full_drop_table;")
 }
 
-// TestCancelDropIndex tests cancel ddl job which type is drop primary key.
-func (s *testDBSuite5) TestCancelDropPrimaryKey(c *C) {
-	idxName := "primary"
-	addIdxSQL := "alter table t add primary key idx_c2 (c2);"
-	dropIdxSQL := "alter table t drop primary key;"
-	testCancelDropIndex(c, s.store, s.dom.DDL(), idxName, addIdxSQL, dropIdxSQL, s.dom)
-}
-
-// TestCancelDropIndex tests cancel ddl job which type is drop index.
-func (s *testDBSuite5) TestCancelDropIndex(c *C) {
-	idxName := "idx_c2"
-	addIdxSQL := "alter table t add index idx_c2 (c2);"
-	dropIdxSQL := "alter table t drop index idx_c2;"
-	testCancelDropIndex(c, s.store, s.dom.DDL(), idxName, addIdxSQL, dropIdxSQL, s.dom)
-}
-
-// testCancelDropIndex tests cancel ddl job which type is drop index.
-func testCancelDropIndex(c *C, store kv.Storage, d ddl.DDL, idxName, addIdxSQL, dropIdxSQL string, dom *domain.Domain) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(c1 int, c2 int)")
-	defer tk.MustExec("drop table t;")
-	for i := 0; i < 5; i++ {
-		tk.MustExec("insert into t values (?, ?)", i, i)
-	}
-	testCases := []struct {
-		needAddIndex   bool
-		jobState       model.JobState
-		JobSchemaState model.SchemaState
-		cancelSucc     bool
-	}{
-		// model.JobStateNone means the jobs is canceled before the first run.
-		// if we cancel successfully, we need to set needAddIndex to false in the next test case. Otherwise, set needAddIndex to true.
-		{true, model.JobStateNone, model.StateNone, true},
-		{false, model.JobStateRunning, model.StateWriteOnly, false},
-		{true, model.JobStateRunning, model.StateDeleteOnly, false},
-		{true, model.JobStateRunning, model.StateDeleteReorganization, false},
-	}
-	var checkErr error
-	hook := &ddl.TestDDLCallback{Do: dom}
-	var jobID int64
-	testCase := &testCases[0]
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if (job.Type == model.ActionDropIndex || job.Type == model.ActionDropPrimaryKey) &&
-			job.State == testCase.jobState && job.SchemaState == testCase.JobSchemaState {
-			jobID = job.ID
-			jobIDs := []int64{job.ID}
-			hookCtx := mock.NewContext()
-			hookCtx.Store = store
-			err := hookCtx.NewTxn(context.TODO())
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			txn, err := hookCtx.Txn(true)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-
-			errs, err := admin.CancelJobs(txn, jobIDs)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			if errs[0] != nil {
-				checkErr = errors.Trace(errs[0])
-				return
-			}
-			checkErr = txn.Commit(context.Background())
-		}
-	}
-	originalHook := d.GetHook()
-	d.SetHook(hook)
-	ctx := tk.Se.(sessionctx.Context)
-	for i := range testCases {
-		testCase = &testCases[i]
-		if testCase.needAddIndex {
-			tk.MustExec(addIdxSQL)
-		}
-		rs, err := tk.Exec(dropIdxSQL)
-		if rs != nil {
-			rs.Close()
-		}
-		t := testGetTableByName(c, ctx, "test_db", "t")
-		indexInfo := t.Meta().FindIndexByName(idxName)
-		if testCase.cancelSucc {
-			c.Assert(checkErr, IsNil)
-			c.Assert(err, NotNil)
-			c.Assert(err.Error(), Equals, "[ddl:8214]Cancelled DDL job")
-			c.Assert(indexInfo, NotNil)
-			c.Assert(indexInfo.State, Equals, model.StatePublic)
-		} else {
-			err1 := admin.ErrCannotCancelDDLJob.GenWithStackByArgs(jobID)
-			c.Assert(err, IsNil)
-			c.Assert(checkErr, NotNil)
-			c.Assert(checkErr.Error(), Equals, err1.Error())
-			c.Assert(indexInfo, IsNil)
-		}
-	}
-	d.SetHook(originalHook)
-	tk.MustExec(addIdxSQL)
-	tk.MustExec(dropIdxSQL)
-}
-
-// TestCancelTruncateTable tests cancel ddl job which type is truncate table.
-func (s *testDBSuite5) TestCancelTruncateTable(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test_db")
-	tk.MustExec("create database if not exists test_truncate_table")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(c1 int, c2 int)")
-	defer tk.MustExec("drop table t;")
-	var checkErr error
-	hook := &ddl.TestDDLCallback{Do: s.dom}
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if job.Type == model.ActionTruncateTable && job.State == model.JobStateNone {
-			jobIDs := []int64{job.ID}
-			hookCtx := mock.NewContext()
-			hookCtx.Store = s.store
-			err := hookCtx.NewTxn(context.Background())
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			txn, err := hookCtx.Txn(true)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			errs, err := admin.CancelJobs(txn, jobIDs)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			if errs[0] != nil {
-				checkErr = errors.Trace(errs[0])
-				return
-			}
-			checkErr = txn.Commit(context.Background())
-		}
-	}
-	originalHook := s.dom.DDL().GetHook()
-	s.dom.DDL().SetHook(hook)
-	_, err := tk.Exec("truncate table t")
-	c.Assert(checkErr, IsNil)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "[ddl:8214]Cancelled DDL job")
-	s.dom.DDL().SetHook(originalHook)
-}
-
-func (s *testDBSuite5) TestParallelDropSchemaAndDropTable(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("create database if not exists test_drop_schema_table")
-	tk.MustExec("use test_drop_schema_table")
-	tk.MustExec("create table t(c1 int, c2 int)")
-	var checkErr error
-	hook := &ddl.TestDDLCallback{Do: s.dom}
-	dbInfo := testGetSchemaByName(c, tk.Se, "test_drop_schema_table")
-	done := false
-	var wg sync.WaitGroup
-	tk2 := testkit.NewTestKit(c, s.store)
-	tk2.MustExec("use test_drop_schema_table")
-	hook.OnJobUpdatedExported = func(job *model.Job) {
-		if job.Type == model.ActionDropSchema && job.State == model.JobStateRunning &&
-			job.SchemaState == model.StateWriteOnly && job.SchemaID == dbInfo.ID && done == false {
-			wg.Add(1)
-			done = true
-			go func() {
-				_, checkErr = tk2.Exec("drop table t")
-				wg.Done()
-			}()
-			time.Sleep(5 * time.Millisecond)
-		}
-	}
-	originalHook := s.dom.DDL().GetHook()
-	s.dom.DDL().SetHook(hook)
-	tk.MustExec("drop database test_drop_schema_table")
-	s.dom.DDL().SetHook(originalHook)
-	wg.Wait()
-	c.Assert(done, IsTrue)
-	c.Assert(checkErr, NotNil)
-	// There are two possible assert result because:
-	// 1: If drop-database is finished before drop-table being put into the ddl job queue, it will return "unknown table" error directly in the previous check.
-	// 2: If drop-table has passed the previous check and been put into the ddl job queue, then drop-database finished, it will return schema change error.
-	assertRes := checkErr.Error() == "[domain:8028]Information schema is changed during the execution of the"+
-		" statement(for example, table definition may be updated by other DDL ran in parallel). "+
-		"If you see this error often, try increasing `tidb_max_delta_schema_count`. [try again later]" ||
-		checkErr.Error() == "[schema:1051]Unknown table 'test_drop_schema_table.t'"
-
-	c.Assert(assertRes, Equals, true)
-
-	// Below behaviour is use to mock query `curl "http://$IP:10080/tiflash/replica"`
-	fn := func(jobs []*model.Job) (bool, error) {
-		return executor.GetDropOrTruncateTableInfoFromJobs(jobs, 0, s.dom, func(job *model.Job, info *model.TableInfo) (bool, error) {
-			return false, nil
-		})
-	}
-	err := tk.Se.NewTxn(context.Background())
-	c.Assert(err, IsNil)
-	txn, err := tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	err = admin.IterHistoryDDLJobs(txn, fn)
-	c.Assert(err, IsNil)
-}
-
-func (s *testDBSuite5) TestAddMultiColumnsIndex(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-
-	tk.MustExec("drop database if exists tidb;")
-	tk.MustExec("create database tidb;")
-	tk.MustExec("use tidb;")
-	tk.MustExec("create table tidb.test (a int auto_increment primary key, b int);")
-	tk.MustExec("insert tidb.test values (1, 1);")
-	tk.MustExec("update tidb.test set b = b + 1 where a = 1;")
-	tk.MustExec("insert into tidb.test values (2, 2);")
-	// Test that the b value is nil.
-	tk.MustExec("insert into tidb.test (a) values (3);")
-	tk.MustExec("insert into tidb.test values (4, 4);")
-	// Test that the b value is nil again.
-	tk.MustExec("insert into tidb.test (a) values (5);")
-	tk.MustExec("insert tidb.test values (6, 6);")
-	tk.MustExec("alter table tidb.test add index idx1 (a, b);")
-	tk.MustExec("admin check table test")
-}
-
 func testGetIndexID(t *testing.T, ctx sessionctx.Context, dbName, tblName, idxName string) int64 {
 	is := domain.GetDomain(ctx).InfoSchema()
 	tt, err := is.TableByName(model.NewCIStr(dbName), model.NewCIStr(tblName))
@@ -603,152 +332,6 @@ func testGetIndexID(t *testing.T, ctx sessionctx.Context, dbName, tblName, idxNa
 	}
 	require.FailNowf(t, "index %s not found(db: %s, tbl: %s)", idxName, dbName, tblName)
 	return -1
-}
-
-type testDDLJobIDCallback struct {
-	ddl.Callback
-	jobID int64
-}
-
-func (t *testDDLJobIDCallback) OnJobUpdated(job *model.Job) {
-	if t.jobID == 0 {
-		t.jobID = job.ID
-	}
-	if t.Callback != nil {
-		t.Callback.OnJobUpdated(job)
-	}
-}
-
-func wrapJobIDExtCallback(oldCallback ddl.Callback) *testDDLJobIDCallback {
-	return &testDDLJobIDCallback{
-		Callback: oldCallback,
-		jobID:    0,
-	}
-}
-
-func setupJobIDExtCallback(ctx sessionctx.Context) (jobExt *testDDLJobIDCallback, tearDown func()) {
-	dom := domain.GetDomain(ctx)
-	originHook := dom.DDL().GetHook()
-	jobIDExt := wrapJobIDExtCallback(originHook)
-	dom.DDL().SetHook(jobIDExt)
-	return jobIDExt, func() {
-		dom.DDL().SetHook(originHook)
-	}
-}
-
-func checkDelRangeAdded(tk *ntestkit.TestKit, jobID int64, elemID int64) {
-	query := `select sum(cnt) from
-	(select count(1) cnt from mysql.gc_delete_range where job_id = ? and element_id = ? union
-	select count(1) cnt from mysql.gc_delete_range_done where job_id = ? and element_id = ?) as gdr;`
-	tk.MustQuery(query, jobID, elemID, jobID, elemID).Check(testkit.Rows("1"))
-}
-
-func (s *testDBSuite5) TestAlterPrimaryKey(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("create table test_add_pk(a int, b int unsigned , c varchar(255) default 'abc', d int as (a+b), e int as (a+1) stored, index idx(b))")
-	defer tk.MustExec("drop table test_add_pk")
-
-	// for generated columns
-	tk.MustGetErrCode("alter table test_add_pk add primary key(d);", errno.ErrUnsupportedOnGeneratedColumn)
-	// The primary key name is the same as the existing index name.
-	tk.MustExec("alter table test_add_pk add primary key idx(e)")
-	tk.MustExec("drop index `primary` on test_add_pk")
-
-	// for describing table
-	tk.MustExec("create table test_add_pk1(a int, index idx(a))")
-	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),YES,MUL,<nil>,`))
-	tk.MustExec("alter table test_add_pk1 add primary key idx(a)")
-	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),NO,PRI,<nil>,`))
-	tk.MustExec("alter table test_add_pk1 drop primary key")
-	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),NO,MUL,<nil>,`))
-	tk.MustExec("create table test_add_pk2(a int, b int, index idx(a))")
-	tk.MustExec("alter table test_add_pk2 add primary key idx(a, b)")
-	tk.MustQuery("desc test_add_pk2").Check(testutil.RowsWithSep(",", ""+
-		"a int(11) NO PRI <nil> ]\n"+
-		"[b int(11) NO PRI <nil> "))
-	tk.MustQuery("show create table test_add_pk2").Check(testutil.RowsWithSep("|", ""+
-		"test_add_pk2 CREATE TABLE `test_add_pk2` (\n"+
-		"  `a` int(11) NOT NULL,\n"+
-		"  `b` int(11) NOT NULL,\n"+
-		"  KEY `idx` (`a`),\n"+
-		"  PRIMARY KEY (`a`,`b`) /*T![clustered_index] NONCLUSTERED */\n"+
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	tk.MustExec("alter table test_add_pk2 drop primary key")
-	tk.MustQuery("desc test_add_pk2").Check(testutil.RowsWithSep(",", ""+
-		"a int(11) NO MUL <nil> ]\n"+
-		"[b int(11) NO  <nil> "))
-
-	// Check if the primary key exists before checking the table's pkIsHandle.
-	tk.MustGetErrCode("alter table test_add_pk drop primary key", errno.ErrCantDropFieldOrKey)
-
-	// for the limit of name
-	validName := strings.Repeat("a", mysql.MaxIndexIdentifierLen)
-	invalidName := strings.Repeat("b", mysql.MaxIndexIdentifierLen+1)
-	tk.MustGetErrCode("alter table test_add_pk add primary key "+invalidName+"(a)", errno.ErrTooLongIdent)
-	// for valid name
-	tk.MustExec("alter table test_add_pk add primary key " + validName + "(a)")
-	// for multiple primary key
-	tk.MustGetErrCode("alter table test_add_pk add primary key (a)", errno.ErrMultiplePriKey)
-	tk.MustExec("alter table test_add_pk drop primary key")
-	// for not existing primary key
-	tk.MustGetErrCode("alter table test_add_pk drop primary key", errno.ErrCantDropFieldOrKey)
-	tk.MustGetErrCode("drop index `primary` on test_add_pk", errno.ErrCantDropFieldOrKey)
-
-	// for too many key parts specified
-	tk.MustGetErrCode("alter table test_add_pk add primary key idx_test(f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f11,f12,f13,f14,f15,f16,f17);",
-		errno.ErrTooManyKeyParts)
-
-	// for the limit of comment's length
-	validComment := "'" + strings.Repeat("a", ddl.MaxCommentLength) + "'"
-	invalidComment := "'" + strings.Repeat("b", ddl.MaxCommentLength+1) + "'"
-	tk.MustGetErrCode("alter table test_add_pk add primary key(a) comment "+invalidComment, errno.ErrTooLongIndexComment)
-	// for empty sql_mode
-	r := tk.MustQuery("select @@sql_mode")
-	sqlMode := r.Rows()[0][0].(string)
-	tk.MustExec("set @@sql_mode=''")
-	tk.MustExec("alter table test_add_pk add primary key(a) comment " + invalidComment)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(1))
-	tk.MustQuery("show warnings").Check(testutil.RowsWithSep("|", "Warning|1688|Comment for index 'PRIMARY' is too long (max = 1024)"))
-	tk.MustExec("set @@sql_mode= '" + sqlMode + "'")
-	tk.MustExec("alter table test_add_pk drop primary key")
-	// for valid comment
-	tk.MustExec("alter table test_add_pk add primary key(a, b, c) comment " + validComment)
-	ctx := tk.Se.(sessionctx.Context)
-	c.Assert(ctx.NewTxn(context.Background()), IsNil)
-	t := testGetTableByName(c, ctx, "test", "test_add_pk")
-	col1Flag := t.Cols()[0].Flag
-	col2Flag := t.Cols()[1].Flag
-	col3Flag := t.Cols()[2].Flag
-	c.Assert(mysql.HasNotNullFlag(col1Flag) && !mysql.HasPreventNullInsertFlag(col1Flag), IsTrue)
-	c.Assert(mysql.HasNotNullFlag(col2Flag) && !mysql.HasPreventNullInsertFlag(col2Flag) && mysql.HasUnsignedFlag(col2Flag), IsTrue)
-	c.Assert(mysql.HasNotNullFlag(col3Flag) && !mysql.HasPreventNullInsertFlag(col3Flag) && !mysql.HasNoDefaultValueFlag(col3Flag), IsTrue)
-	tk.MustExec("alter table test_add_pk drop primary key")
-
-	// for null values in primary key
-	tk.MustExec("drop table test_add_pk")
-	tk.MustExec("create table test_add_pk(a int, b int unsigned , c varchar(255) default 'abc', index idx(b))")
-	tk.MustExec("insert into test_add_pk set a = 0, b = 0, c = 0")
-	tk.MustExec("insert into test_add_pk set a = 1")
-	tk.MustGetErrCode("alter table test_add_pk add primary key (b)", errno.ErrInvalidUseOfNull)
-	tk.MustExec("insert into test_add_pk set a = 2, b = 2")
-	tk.MustGetErrCode("alter table test_add_pk add primary key (a, b)", errno.ErrInvalidUseOfNull)
-	tk.MustExec("insert into test_add_pk set a = 3, c = 3")
-	tk.MustGetErrCode("alter table test_add_pk add primary key (c, b, a)", errno.ErrInvalidUseOfNull)
-}
-
-func (s *testDBSuite5) TestCreateIndexType(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-	sql := `CREATE TABLE test_index (
-		price int(5) DEFAULT '0' NOT NULL,
-		area varchar(40) DEFAULT '' NOT NULL,
-		type varchar(40) DEFAULT '' NOT NULL,
-		transityes set('a','b'),
-		shopsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
-		schoolsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
-		petsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
-		KEY price (price,area,type,transityes,shopsyes,schoolsyes,petsyes));`
-	tk.MustExec(sql)
 }
 
 func oldBackgroundExec(s kv.Storage, sql string, done chan error) {
@@ -1010,241 +593,6 @@ func (s *testSerialDBSuite) TestRebaseAutoID(c *C) {
 
 	tk.MustExec("create table tidb.test2 (a int);")
 	tk.MustGetErrCode("alter table tidb.test2 add column b int auto_increment key, auto_increment=10;", errno.ErrUnsupportedDDLOperation)
-}
-
-func (s *testDBSuite5) TestCheckColumnDefaultValue(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists text_default_text;")
-	tk.MustGetErrCode("create table text_default_text(c1 text not null default '');", errno.ErrBlobCantHaveDefault)
-	tk.MustGetErrCode("create table text_default_text(c1 text not null default 'scds');", errno.ErrBlobCantHaveDefault)
-
-	tk.MustExec("drop table if exists text_default_json;")
-	tk.MustGetErrCode("create table text_default_json(c1 json not null default '');", errno.ErrBlobCantHaveDefault)
-	tk.MustGetErrCode("create table text_default_json(c1 json not null default 'dfew555');", errno.ErrBlobCantHaveDefault)
-
-	tk.MustExec("drop table if exists text_default_blob;")
-	tk.MustGetErrCode("create table text_default_blob(c1 blob not null default '');", errno.ErrBlobCantHaveDefault)
-	tk.MustGetErrCode("create table text_default_blob(c1 blob not null default 'scds54');", errno.ErrBlobCantHaveDefault)
-
-	tk.MustExec("set sql_mode='';")
-	tk.MustExec("create table text_default_text(c1 text not null default '');")
-	tk.MustQuery(`show create table text_default_text`).Check(testutil.RowsWithSep("|",
-		"text_default_text CREATE TABLE `text_default_text` (\n"+
-			"  `c1` text NOT NULL\n"+
-			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
-	))
-	ctx := tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	tblInfo, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("text_default_text"))
-	c.Assert(err, IsNil)
-	c.Assert(tblInfo.Meta().Columns[0].DefaultValue, Equals, "")
-
-	tk.MustExec("create table text_default_blob(c1 blob not null default '');")
-	tk.MustQuery(`show create table text_default_blob`).Check(testutil.RowsWithSep("|",
-		"text_default_blob CREATE TABLE `text_default_blob` (\n"+
-			"  `c1` blob NOT NULL\n"+
-			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
-	))
-	is = domain.GetDomain(ctx).InfoSchema()
-	tblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("text_default_blob"))
-	c.Assert(err, IsNil)
-	c.Assert(tblInfo.Meta().Columns[0].DefaultValue, Equals, "")
-
-	tk.MustExec("create table text_default_json(c1 json not null default '');")
-	tk.MustQuery(`show create table text_default_json`).Check(testutil.RowsWithSep("|",
-		"text_default_json CREATE TABLE `text_default_json` (\n"+
-			"  `c1` json NOT NULL DEFAULT 'null'\n"+
-			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
-	))
-	is = domain.GetDomain(ctx).InfoSchema()
-	tblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("text_default_json"))
-	c.Assert(err, IsNil)
-	c.Assert(tblInfo.Meta().Columns[0].DefaultValue, Equals, `null`)
-}
-
-func (s *testDBSuite5) TestCheckConvertToCharacter(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	defer tk.MustExec("drop table t")
-	tk.MustExec("create table t(a varchar(10) charset binary);")
-	ctx := tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	t, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("t"))
-	c.Assert(err, IsNil)
-	tk.MustGetErrCode("alter table t modify column a varchar(10) charset utf8 collate utf8_bin", errno.ErrUnsupportedDDLOperation)
-	tk.MustGetErrCode("alter table t modify column a varchar(10) charset utf8mb4 collate utf8mb4_bin", errno.ErrUnsupportedDDLOperation)
-	tk.MustGetErrCode("alter table t modify column a varchar(10) charset latin1 collate latin1_bin", errno.ErrUnsupportedDDLOperation)
-	c.Assert(t.Cols()[0].Charset, Equals, "binary")
-}
-
-func (s *testDBSuite5) TestModifyGeneratedColumn(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("create database if not exists test;")
-	tk.MustExec("use test")
-	modIdxColErrMsg := "[ddl:3106]'modifying an indexed column' is not supported for generated columns."
-	modStoredColErrMsg := "[ddl:3106]'modifying a stored column' is not supported for generated columns."
-
-	// Modify column with single-col-index.
-	tk.MustExec("drop table if exists t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1), index idx(b));")
-	tk.MustExec("insert into t1 set a=1;")
-	_, err := tk.Exec("alter table t1 modify column b int as (a+2);")
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, modIdxColErrMsg)
-	tk.MustExec("drop index idx on t1;")
-	tk.MustExec("alter table t1 modify b int as (a+2);")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("1 3"))
-
-	// Modify column with multi-col-index.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1), index idx(a, b));")
-	tk.MustExec("insert into t1 set a=1;")
-	_, err = tk.Exec("alter table t1 modify column b int as (a+2);")
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, modIdxColErrMsg)
-	tk.MustExec("drop index idx on t1;")
-	tk.MustExec("alter table t1 modify b int as (a+2);")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("1 3"))
-
-	// Modify column with stored status to a different expression.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1) stored);")
-	tk.MustExec("insert into t1 set a=1;")
-	_, err = tk.Exec("alter table t1 modify column b int as (a+2) stored;")
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, modStoredColErrMsg)
-
-	// Modify column with stored status to the same expression.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1) stored);")
-	tk.MustExec("insert into t1 set a=1;")
-	tk.MustExec("alter table t1 modify column b bigint as (a+1) stored;")
-	tk.MustExec("alter table t1 modify column b bigint as (a + 1) stored;")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("1 2"))
-
-	// Modify column with index to the same expression.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1), index idx(b));")
-	tk.MustExec("insert into t1 set a=1;")
-	tk.MustExec("alter table t1 modify column b bigint as (a+1);")
-	tk.MustExec("alter table t1 modify column b bigint as (a + 1);")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("1 2"))
-
-	// Modify column from non-generated to stored generated.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int);")
-	_, err = tk.Exec("alter table t1 modify column b bigint as (a+1) stored;")
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, modStoredColErrMsg)
-
-	// Modify column from stored generated to non-generated.
-	tk.MustExec("drop table t1;")
-	tk.MustExec("create table t1 (a int, b int as (a+1) stored);")
-	tk.MustExec("insert into t1 set a=1;")
-	tk.MustExec("alter table t1 modify column b int;")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("1 2"))
-}
-
-func (s *testDBSuite5) TestDefaultSQLFunction(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("create database if not exists test;")
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists t1, t2, t3, t4;")
-
-	// For issue #13189
-	// Use `DEFAULT()` in `INSERT` / `INSERT ON DUPLICATE KEY UPDATE` statement
-	tk.MustExec("create table t1 (a int primary key, b int default 20, c int default 30, d int default 40);")
-	tk.MustExec("SET @@time_zone = '+00:00'")
-	defer tk.MustExec("SET @@time_zone = DEFAULT")
-	tk.MustQuery("SELECT @@time_zone").Check(testkit.Rows("+00:00"))
-	tk.MustExec("create table t2 (a int primary key, b timestamp DEFAULT CURRENT_TIMESTAMP, c timestamp DEFAULT '2000-01-01 00:00:00')")
-	tk.MustExec("insert into t1 set a = 1, b = default(c);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40"))
-	tk.MustExec("insert into t1 set a = 2, b = default(c), c = default(d), d = default(b);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40", "2 30 40 20"))
-	tk.MustExec("insert into t1 values (2, 3, 4, 5) on duplicate key update b = default(d), c = default(b);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40", "2 40 20 20"))
-	tk.MustExec("delete from t1")
-	tk.MustExec("insert into t1 set a = default(b) + default(c) - default(d)")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 20 30 40"))
-	tk.MustExec("set @@timestamp = 1321009871")
-	defer tk.MustExec("set @@timestamp = DEFAULT")
-	tk.MustQuery("SELECT NOW()").Check(testkit.Rows("2011-11-11 11:11:11"))
-	tk.MustExec("insert into t2 set a = 1, b = default(c)")
-	tk.MustExec("insert into t2 set a = 2, c = default(b)")
-	tk.MustGetErrCode("insert into t2 set a = 3, b = default(a)", errno.ErrNoDefaultForField)
-	tk.MustExec("insert into t2 set a = 4, b = default(b), c = default(c)")
-	tk.MustExec("insert into t2 set a = 5, b = default, c = default")
-	tk.MustExec("insert into t2 set a = 6")
-	tk.MustQuery("select * from t2").Sort().Check(testkit.Rows(
-		"1 2000-01-01 00:00:00 2000-01-01 00:00:00",
-		"2 2011-11-11 11:11:11 2011-11-11 11:11:11",
-		"4 2011-11-11 11:11:11 2000-01-01 00:00:00",
-		"5 2011-11-11 11:11:11 2000-01-01 00:00:00",
-		"6 2011-11-11 11:11:11 2000-01-01 00:00:00"))
-	// Use `DEFAULT()` in `UPDATE` statement
-	tk.MustExec("delete from t1;")
-	tk.MustExec("insert into t1 value (1, 2, 3, 4);")
-	tk.MustExec("update t1 set a = 1, c = default(b);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 2 20 4"))
-	tk.MustExec("insert into t1 value (2, 2, 3, 4);")
-	tk.MustExec("update t1 set c = default(b), b = default(c) where a = 2;")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 2 20 4", "2 30 20 4"))
-	tk.MustExec("delete from t1")
-	tk.MustExec("insert into t1 set a = 10")
-	tk.MustExec("update t1 set a = 10, b = default(c) + default(d)")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40"))
-	tk.MustExec("set @@timestamp = 1671747742")
-	tk.MustExec("update t2 set b = default(c) WHERE a = 6")
-	tk.MustExec("update t2 set c = default(b) WHERE a = 5")
-	tk.MustGetErrCode("update t2 set b = default(a) WHERE a = 4", errno.ErrNoDefaultForField)
-	tk.MustExec("update t2 set b = default(b), c = default(c) WHERE a = 4")
-	// Non existing row!
-	tk.MustExec("update t2 set b = default(b), c = default(c) WHERE a = 3")
-	tk.MustExec("update t2 set b = default, c = default WHERE a = 2")
-	tk.MustExec("update t2 set b = default(b) WHERE a = 1")
-	tk.MustQuery("select * from t2;").Sort().Check(testkit.Rows(
-		"1 2022-12-22 22:22:22 2000-01-01 00:00:00",
-		"2 2022-12-22 22:22:22 2000-01-01 00:00:00",
-		"4 2022-12-22 22:22:22 2000-01-01 00:00:00",
-		"5 2011-11-11 11:11:11 2022-12-22 22:22:22",
-		"6 2000-01-01 00:00:00 2000-01-01 00:00:00"))
-	// Use `DEFAULT()` in `REPLACE` statement
-	tk.MustExec("delete from t1;")
-	tk.MustExec("insert into t1 value (1, 2, 3, 4);")
-	tk.MustExec("replace into t1 set a = 1, c = default(b);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 20 20 40"))
-	tk.MustExec("insert into t1 value (2, 2, 3, 4);")
-	tk.MustExec("replace into t1 set a = 2, d = default(b), c = default(d);")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 20 20 40", "2 20 40 20"))
-	tk.MustExec("delete from t1")
-	tk.MustExec("insert into t1 set a = 10, c = 3")
-	tk.MustExec("replace into t1 set a = 10, b = default(c) + default(d)")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40"))
-	tk.MustExec("replace into t1 set a = 20, d = default(c) + default(b)")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40", "20 20 30 50"))
-
-	// Use `DEFAULT()` in expression of generate columns, issue #12471
-	tk.MustExec("DROP TABLE t2")
-	tk.MustExec("create table t2(a int default 9, b int as (1 + default(a)));")
-	tk.MustExec("insert into t2 values(1, default);")
-	tk.MustExec("insert into t2 values(2, default(b))")
-	tk.MustQuery("select * from t2").Sort().Check(testkit.Rows("1 10", "2 10"))
-
-	// Use `DEFAULT()` with subquery, issue #13390
-	tk.MustExec("create table t3(f1 int default 11);")
-	tk.MustExec("insert into t3 value ();")
-	tk.MustQuery("select default(f1) from (select * from t3) t1;").Check(testkit.Rows("11"))
-	tk.MustQuery("select default(f1) from (select * from (select * from t3) t1 ) t1;").Check(testkit.Rows("11"))
-
-	tk.MustExec("create table t4(a int default 4);")
-	tk.MustExec("insert into t4 value (2);")
-	tk.MustQuery("select default(c) from (select b as c from (select a as b from t4) t3) t2;").Check(testkit.Rows("4"))
-	tk.MustGetErrCode("select default(a) from (select a from (select 1 as a) t4) t4;", errno.ErrNoDefaultForField)
-
-	tk.MustExec("drop table t1, t2, t3, t4;")
 }
 
 func (s *testSerialDBSuite) TestProcessColumnFlags(c *C) {
@@ -1562,17 +910,6 @@ func (s *testSerialDBSuite) TestSkipSchemaChecker(c *C) {
 	tk2.MustExec("alter table t1 add column b int;")
 	_, err = tk.Exec("commit")
 	c.Assert(terror.ErrorEqual(domain.ErrInfoSchemaChanged, err), IsTrue)
-}
-
-func (s *testDBSuite5) TestAlterCheck(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use " + s.schemaName)
-	tk.MustExec("drop table if exists alter_check")
-	tk.MustExec("create table alter_check (pk int primary key)")
-	defer tk.MustExec("drop table if exists alter_check")
-	tk.MustExec("alter table alter_check alter check crcn ENFORCED")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(1))
-	tk.MustQuery("show warnings").Check(testutil.RowsWithSep("|", "Warning|8231|ALTER CHECK is not supported"))
 }
 
 func (s *testSerialDBSuite) TestDDLJobErrorCount(c *C) {
