+++ ddl/db_test.go
@@ -15,21 +15,29 @@
 package ddl_test
 
 import (
+	"context"
 	"fmt"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
+	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/ddl"
 	ddlutil "github.com/pingcap/tidb/ddl/util"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/errno"
+	"github.com/pingcap/tidb/executor"
+	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/testkit/external"
 	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/admin"
+	"github.com/pingcap/tidb/util/mock"
+	"github.com/pingcap/tidb/util/testutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -727,4 +735,432 @@ func TestCreateTableWithDecimalWithDoubleZero(t *testing.T) {
 	tk.MustExec("create table tt(d int)")
 	tk.MustExec("alter table tt change column d d decimal(0, 0)")
 	checkType("test", "tt", "d")
+}
+
+func TestAlterCheck(t *testing.T) {
+	store, clean := testkit.CreateMockStoreWithSchemaLease(t, dbTestLease)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table alter_check (pk int primary key)")
+	tk.MustExec("alter table alter_check alter check crcn ENFORCED")
+	require.Equal(t, uint16(1), tk.Session().GetSessionVars().StmtCtx.WarningCount())
+	tk.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Warning|8231|ALTER CHECK is not supported"))
+}
+
+func TestDefaultSQLFunction(t *testing.T) {
+	store, clean := testkit.CreateMockStoreWithSchemaLease(t, dbTestLease)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+
+	// For issue #13189
+	// Use `DEFAULT()` in `INSERT` / `INSERT ON DUPLICATE KEY UPDATE` statement
+	tk.MustExec("create table t1 (a int primary key, b int default 20, c int default 30, d int default 40);")
+	tk.MustExec("SET @@time_zone = '+00:00'")
+	defer tk.MustExec("SET @@time_zone = DEFAULT")
+	tk.MustQuery("SELECT @@time_zone").Check(testkit.Rows("+00:00"))
+	tk.MustExec("create table t2 (a int primary key, b timestamp DEFAULT CURRENT_TIMESTAMP, c timestamp DEFAULT '2000-01-01 00:00:00')")
+	tk.MustExec("insert into t1 set a = 1, b = default(c);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40"))
+	tk.MustExec("insert into t1 set a = 2, b = default(c), c = default(d), d = default(b);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40", "2 30 40 20"))
+	tk.MustExec("insert into t1 values (2, 3, 4, 5) on duplicate key update b = default(d), c = default(b);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 30 30 40", "2 40 20 20"))
+	tk.MustExec("delete from t1")
+	tk.MustExec("insert into t1 set a = default(b) + default(c) - default(d)")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 20 30 40"))
+	tk.MustExec("set @@timestamp = 1321009871")
+	defer tk.MustExec("set @@timestamp = DEFAULT")
+	tk.MustQuery("SELECT NOW()").Check(testkit.Rows("2011-11-11 11:11:11"))
+	tk.MustExec("insert into t2 set a = 1, b = default(c)")
+	tk.MustExec("insert into t2 set a = 2, c = default(b)")
+	tk.MustGetErrCode("insert into t2 set a = 3, b = default(a)", errno.ErrNoDefaultForField)
+	tk.MustExec("insert into t2 set a = 4, b = default(b), c = default(c)")
+	tk.MustExec("insert into t2 set a = 5, b = default, c = default")
+	tk.MustExec("insert into t2 set a = 6")
+	tk.MustQuery("select * from t2").Sort().Check(testkit.Rows(
+		"1 2000-01-01 00:00:00 2000-01-01 00:00:00",
+		"2 2011-11-11 11:11:11 2011-11-11 11:11:11",
+		"4 2011-11-11 11:11:11 2000-01-01 00:00:00",
+		"5 2011-11-11 11:11:11 2000-01-01 00:00:00",
+		"6 2011-11-11 11:11:11 2000-01-01 00:00:00"))
+	// Use `DEFAULT()` in `UPDATE` statement
+	tk.MustExec("delete from t1;")
+	tk.MustExec("insert into t1 value (1, 2, 3, 4);")
+	tk.MustExec("update t1 set a = 1, c = default(b);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 2 20 4"))
+	tk.MustExec("insert into t1 value (2, 2, 3, 4);")
+	tk.MustExec("update t1 set c = default(b), b = default(c) where a = 2;")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 2 20 4", "2 30 20 4"))
+	tk.MustExec("delete from t1")
+	tk.MustExec("insert into t1 set a = 10")
+	tk.MustExec("update t1 set a = 10, b = default(c) + default(d)")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40"))
+	tk.MustExec("set @@timestamp = 1671747742")
+	tk.MustExec("update t2 set b = default(c) WHERE a = 6")
+	tk.MustExec("update t2 set c = default(b) WHERE a = 5")
+	tk.MustGetErrCode("update t2 set b = default(a) WHERE a = 4", errno.ErrNoDefaultForField)
+	tk.MustExec("update t2 set b = default(b), c = default(c) WHERE a = 4")
+	// Non existing row!
+	tk.MustExec("update t2 set b = default(b), c = default(c) WHERE a = 3")
+	tk.MustExec("update t2 set b = default, c = default WHERE a = 2")
+	tk.MustExec("update t2 set b = default(b) WHERE a = 1")
+	tk.MustQuery("select * from t2;").Sort().Check(testkit.Rows(
+		"1 2022-12-22 22:22:22 2000-01-01 00:00:00",
+		"2 2022-12-22 22:22:22 2000-01-01 00:00:00",
+		"4 2022-12-22 22:22:22 2000-01-01 00:00:00",
+		"5 2011-11-11 11:11:11 2022-12-22 22:22:22",
+		"6 2000-01-01 00:00:00 2000-01-01 00:00:00"))
+	// Use `DEFAULT()` in `REPLACE` statement
+	tk.MustExec("delete from t1;")
+	tk.MustExec("insert into t1 value (1, 2, 3, 4);")
+	tk.MustExec("replace into t1 set a = 1, c = default(b);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 20 20 40"))
+	tk.MustExec("insert into t1 value (2, 2, 3, 4);")
+	tk.MustExec("replace into t1 set a = 2, d = default(b), c = default(d);")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("1 20 20 40", "2 20 40 20"))
+	tk.MustExec("delete from t1")
+	tk.MustExec("insert into t1 set a = 10, c = 3")
+	tk.MustExec("replace into t1 set a = 10, b = default(c) + default(d)")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40"))
+	tk.MustExec("replace into t1 set a = 20, d = default(c) + default(b)")
+	tk.MustQuery("select * from t1;").Check(testkit.Rows("10 70 30 40", "20 20 30 50"))
+
+	// Use `DEFAULT()` in expression of generate columns, issue #12471
+	tk.MustExec("DROP TABLE t2")
+	tk.MustExec("create table t2(a int default 9, b int as (1 + default(a)));")
+	tk.MustExec("insert into t2 values(1, default);")
+	tk.MustExec("insert into t2 values(2, default(b))")
+	tk.MustQuery("select * from t2").Sort().Check(testkit.Rows("1 10", "2 10"))
+
+	// Use `DEFAULT()` with subquery, issue #13390
+	tk.MustExec("create table t3(f1 int default 11);")
+	tk.MustExec("insert into t3 value ();")
+	tk.MustQuery("select default(f1) from (select * from t3) t1;").Check(testkit.Rows("11"))
+	tk.MustQuery("select default(f1) from (select * from (select * from t3) t1 ) t1;").Check(testkit.Rows("11"))
+
+	tk.MustExec("create table t4(a int default 4);")
+	tk.MustExec("insert into t4 value (2);")
+	tk.MustQuery("select default(c) from (select b as c from (select a as b from t4) t3) t2;").Check(testkit.Rows("4"))
+	tk.MustGetErrCode("select default(a) from (select a from (select 1 as a) t4) t4;", errno.ErrNoDefaultForField)
+
+	tk.MustExec("drop table t1, t2, t3, t4;")
+}
+
+func TestCreateIndexType(t *testing.T) {
+	store, clean := testkit.CreateMockStoreWithSchemaLease(t, dbTestLease)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec(`CREATE TABLE test_index (
+		price int(5) DEFAULT '0' NOT NULL,
+		area varchar(40) DEFAULT '' NOT NULL,
+		type varchar(40) DEFAULT '' NOT NULL,
+		transityes set('a','b'),
+		shopsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
+		schoolsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
+		petsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
+		KEY price (price,area,type,transityes,shopsyes,schoolsyes,petsyes));`)
+}
+
+func TestAlterPrimaryKey(t *testing.T) {
+	store, clean := testkit.CreateMockStoreWithSchemaLease(t, dbTestLease)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table test_add_pk(a int, b int unsigned , c varchar(255) default 'abc', d int as (a+b), e int as (a+1) stored, index idx(b))")
+
+	// for generated columns
+	tk.MustGetErrCode("alter table test_add_pk add primary key(d);", errno.ErrUnsupportedOnGeneratedColumn)
+	// The primary key name is the same as the existing index name.
+	tk.MustExec("alter table test_add_pk add primary key idx(e)")
+	tk.MustExec("drop index `primary` on test_add_pk")
+
+	// for describing table
+	tk.MustExec("create table test_add_pk1(a int, index idx(a))")
+	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),YES,MUL,<nil>,`))
+	tk.MustExec("alter table test_add_pk1 add primary key idx(a)")
+	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),NO,PRI,<nil>,`))
+	tk.MustExec("alter table test_add_pk1 drop primary key")
+	tk.MustQuery("desc test_add_pk1").Check(testutil.RowsWithSep(",", `a,int(11),NO,MUL,<nil>,`))
+	tk.MustExec("create table test_add_pk2(a int, b int, index idx(a))")
+	tk.MustExec("alter table test_add_pk2 add primary key idx(a, b)")
+	tk.MustQuery("desc test_add_pk2").Check(testutil.RowsWithSep(",", ""+
+		"a int(11) NO PRI <nil> ]\n"+
+		"[b int(11) NO PRI <nil> "))
+	tk.MustQuery("show create table test_add_pk2").Check(testutil.RowsWithSep("|", ""+
+		"test_add_pk2 CREATE TABLE `test_add_pk2` (\n"+
+		"  `a` int(11) NOT NULL,\n"+
+		"  `b` int(11) NOT NULL,\n"+
+		"  KEY `idx` (`a`),\n"+
+		"  PRIMARY KEY (`a`,`b`) /*T![clustered_index] NONCLUSTERED */\n"+
+		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
+	tk.MustExec("alter table test_add_pk2 drop primary key")
+	tk.MustQuery("desc test_add_pk2").Check(testutil.RowsWithSep(",", ""+
+		"a int(11) NO MUL <nil> ]\n"+
+		"[b int(11) NO  <nil> "))
+
+	// Check if the primary key exists before checking the table's pkIsHandle.
+	tk.MustGetErrCode("alter table test_add_pk drop primary key", errno.ErrCantDropFieldOrKey)
+
+	// for the limit of name
+	validName := strings.Repeat("a", mysql.MaxIndexIdentifierLen)
+	invalidName := strings.Repeat("b", mysql.MaxIndexIdentifierLen+1)
+	tk.MustGetErrCode("alter table test_add_pk add primary key "+invalidName+"(a)", errno.ErrTooLongIdent)
+	// for valid name
+	tk.MustExec("alter table test_add_pk add primary key " + validName + "(a)")
+	// for multiple primary key
+	tk.MustGetErrCode("alter table test_add_pk add primary key (a)", errno.ErrMultiplePriKey)
+	tk.MustExec("alter table test_add_pk drop primary key")
+	// for not existing primary key
+	tk.MustGetErrCode("alter table test_add_pk drop primary key", errno.ErrCantDropFieldOrKey)
+	tk.MustGetErrCode("drop index `primary` on test_add_pk", errno.ErrCantDropFieldOrKey)
+
+	// for too many key parts specified
+	tk.MustGetErrCode("alter table test_add_pk add primary key idx_test(f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f11,f12,f13,f14,f15,f16,f17);",
+		errno.ErrTooManyKeyParts)
+
+	// for the limit of comment's length
+	validComment := "'" + strings.Repeat("a", ddl.MaxCommentLength) + "'"
+	invalidComment := "'" + strings.Repeat("b", ddl.MaxCommentLength+1) + "'"
+	tk.MustGetErrCode("alter table test_add_pk add primary key(a) comment "+invalidComment, errno.ErrTooLongIndexComment)
+	// for empty sql_mode
+	r := tk.MustQuery("select @@sql_mode")
+	sqlMode := r.Rows()[0][0].(string)
+	tk.MustExec("set @@sql_mode=''")
+	tk.MustExec("alter table test_add_pk add primary key(a) comment " + invalidComment)
+	require.Equal(t, uint16(1), tk.Session().GetSessionVars().StmtCtx.WarningCount())
+	tk.MustQuery("show warnings").Check(testutil.RowsWithSep("|", "Warning|1688|Comment for index 'PRIMARY' is too long (max = 1024)"))
+	tk.MustExec("set @@sql_mode= '" + sqlMode + "'")
+	tk.MustExec("alter table test_add_pk drop primary key")
+	// for valid comment
+	tk.MustExec("alter table test_add_pk add primary key(a, b, c) comment " + validComment)
+	require.NoError(t, tk.Session().NewTxn(context.Background()))
+	tbl := external.GetTableByName(t, tk, "test", "test_add_pk")
+	col1Flag := tbl.Cols()[0].Flag
+	col2Flag := tbl.Cols()[1].Flag
+	col3Flag := tbl.Cols()[2].Flag
+	require.True(t, mysql.HasNotNullFlag(col1Flag) && !mysql.HasPreventNullInsertFlag(col1Flag))
+	require.True(t, mysql.HasNotNullFlag(col2Flag) && !mysql.HasPreventNullInsertFlag(col2Flag) && mysql.HasUnsignedFlag(col2Flag))
+	require.True(t, mysql.HasNotNullFlag(col3Flag) && !mysql.HasPreventNullInsertFlag(col3Flag) && !mysql.HasNoDefaultValueFlag(col3Flag))
+	tk.MustExec("alter table test_add_pk drop primary key")
+
+	// for null values in primary key
+	tk.MustExec("drop table test_add_pk")
+	tk.MustExec("create table test_add_pk(a int, b int unsigned , c varchar(255) default 'abc', index idx(b))")
+	tk.MustExec("insert into test_add_pk set a = 0, b = 0, c = 0")
+	tk.MustExec("insert into test_add_pk set a = 1")
+	tk.MustGetErrCode("alter table test_add_pk add primary key (b)", errno.ErrInvalidUseOfNull)
+	tk.MustExec("insert into test_add_pk set a = 2, b = 2")
+	tk.MustGetErrCode("alter table test_add_pk add primary key (a, b)", errno.ErrInvalidUseOfNull)
+	tk.MustExec("insert into test_add_pk set a = 3, c = 3")
+	tk.MustGetErrCode("alter table test_add_pk add primary key (c, b, a)", errno.ErrInvalidUseOfNull)
+}
+
+func TestParallelDropSchemaAndDropTable(t *testing.T) {
+	store, dom, clean := testkit.CreateMockStoreAndDomainWithSchemaLease(t, dbTestLease)
+	defer clean()
+
+	tk1 := testkit.NewTestKit(t, store)
+	tk1.MustExec("create database if not exists test_drop_schema_table")
+	tk1.MustExec("use test_drop_schema_table")
+	tk1.MustExec("create table t(c1 int, c2 int)")
+	var checkErr error
+	hook := &ddl.TestDDLCallback{Do: dom}
+
+	dbInfo := external.GetSchemaByName(t, tk1, "test_drop_schema_table")
+	done := false
+	var wg sync.WaitGroup
+	tk2 := testkit.NewTestKit(t, store)
+	tk2.MustExec("use test_drop_schema_table")
+	hook.OnJobUpdatedExported = func(job *model.Job) {
+		if job.Type == model.ActionDropSchema && job.State == model.JobStateRunning &&
+			job.SchemaState == model.StateWriteOnly && job.SchemaID == dbInfo.ID && done == false {
+			wg.Add(1)
+			done = true
+			go func() {
+				_, checkErr = tk2.Exec("drop table t")
+				wg.Done()
+			}()
+			time.Sleep(5 * time.Millisecond)
+		}
+	}
+	originalHook := dom.DDL().GetHook()
+	dom.DDL().SetHook(hook)
+	tk1.MustExec("drop database test_drop_schema_table")
+	dom.DDL().SetHook(originalHook)
+	wg.Wait()
+	require.True(t, done)
+	require.Error(t, checkErr)
+	// There are two possible assert result because:
+	// 1: If drop-database is finished before drop-table being put into the ddl job queue, it will return "unknown table" error directly in the previous check.
+	// 2: If drop-table has passed the previous check and been put into the ddl job queue, then drop-database finished, it will return schema change error.
+	assertRes := checkErr.Error() == "[domain:8028]Information schema is changed during the execution of the"+
+		" statement(for example, table definition may be updated by other DDL ran in parallel). "+
+		"If you see this error often, try increasing `tidb_max_delta_schema_count`. [try again later]" ||
+		checkErr.Error() == "[schema:1051]Unknown table 'test_drop_schema_table.t'"
+	require.True(t, assertRes)
+
+	// Below behaviour is use to mock query `curl "http://$IP:10080/tiflash/replica"`
+	fn := func(jobs []*model.Job) (bool, error) {
+		return executor.GetDropOrTruncateTableInfoFromJobs(jobs, 0, dom, func(job *model.Job, info *model.TableInfo) (bool, error) {
+			return false, nil
+		})
+	}
+	require.NoError(t, tk1.Session().NewTxn(context.Background()))
+	txn, err := tk1.Session().Txn(true)
+	require.NoError(t, err)
+	require.NoError(t, admin.IterHistoryDDLJobs(txn, fn))
+}
+
+// TestCancelDropIndex tests cancel ddl job which type is drop primary key.
+func TestCancelDropPrimaryKey(t *testing.T) {
+	store, dom, clean := testkit.CreateMockStoreAndDomainWithSchemaLease(t, dbTestLease)
+	defer clean()
+	idxName := "primary"
+	addIdxSQL := "alter table t add primary key idx_c2 (c2);"
+	dropIdxSQL := "alter table t drop primary key;"
+	testCancelDropIndex(t, store, dom.DDL(), idxName, addIdxSQL, dropIdxSQL, dom)
+}
+
+// TestCancelDropIndex tests cancel ddl job which type is drop index.
+func TestCancelDropIndex(t *testing.T) {
+	store, dom, clean := testkit.CreateMockStoreAndDomainWithSchemaLease(t, dbTestLease)
+	defer clean()
+	idxName := "idx_c2"
+	addIdxSQL := "alter table t add index idx_c2 (c2);"
+	dropIdxSQL := "alter table t drop index idx_c2;"
+	testCancelDropIndex(t, store, dom.DDL(), idxName, addIdxSQL, dropIdxSQL, dom)
+}
+
+// testCancelDropIndex tests cancel ddl job which type is drop index.
+func testCancelDropIndex(t *testing.T, store kv.Storage, d ddl.DDL, idxName, addIdxSQL, dropIdxSQL string, dom *domain.Domain) {
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("drop table if exists t")
+	tk.MustExec("create table t(c1 int, c2 int)")
+	defer tk.MustExec("drop table t;")
+	for i := 0; i < 5; i++ {
+		tk.MustExec("insert into t values (?, ?)", i, i)
+	}
+	tests := []struct {
+		needAddIndex   bool
+		jobState       model.JobState
+		JobSchemaState model.SchemaState
+		cancelSucc     bool
+	}{
+		// model.JobStateNone means the jobs is canceled before the first run.
+		// if we cancel successfully, we need to set needAddIndex to false in the next test case. Otherwise, set needAddIndex to true.
+		{true, model.JobStateNone, model.StateNone, true},
+		{false, model.JobStateRunning, model.StateWriteOnly, false},
+		{true, model.JobStateRunning, model.StateDeleteOnly, false},
+		{true, model.JobStateRunning, model.StateDeleteReorganization, false},
+	}
+	var checkErr error
+	hook := &ddl.TestDDLCallback{Do: dom}
+	var jobID int64
+	test := &tests[0]
+	hook.OnJobRunBeforeExported = func(job *model.Job) {
+		if (job.Type == model.ActionDropIndex || job.Type == model.ActionDropPrimaryKey) &&
+			job.State == test.jobState && job.SchemaState == test.JobSchemaState {
+			jobID = job.ID
+			jobIDs := []int64{job.ID}
+			hookCtx := mock.NewContext()
+			hookCtx.Store = store
+			err := hookCtx.NewTxn(context.TODO())
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+			txn, err := hookCtx.Txn(true)
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+
+			errs, err := admin.CancelJobs(txn, jobIDs)
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+			if errs[0] != nil {
+				checkErr = errors.Trace(errs[0])
+				return
+			}
+			checkErr = txn.Commit(context.Background())
+		}
+	}
+	originalHook := d.GetHook()
+	d.SetHook(hook)
+	for i := range tests {
+		test = &tests[i]
+		if test.needAddIndex {
+			tk.MustExec(addIdxSQL)
+		}
+		err := tk.ExecToErr(dropIdxSQL)
+		tbl := external.GetTableByName(t, tk, "test", "t")
+		indexInfo := tbl.Meta().FindIndexByName(idxName)
+		if test.cancelSucc {
+			require.NoError(t, checkErr)
+			require.EqualError(t, err, "[ddl:8214]Cancelled DDL job")
+			require.NotNil(t, indexInfo)
+			require.Equal(t, model.StatePublic, indexInfo.State)
+		} else {
+			err1 := admin.ErrCannotCancelDDLJob.GenWithStackByArgs(jobID)
+			require.NoError(t, err)
+			require.EqualError(t, checkErr, err1.Error())
+			require.Nil(t, indexInfo)
+		}
+	}
+	d.SetHook(originalHook)
+	tk.MustExec(addIdxSQL)
+	tk.MustExec(dropIdxSQL)
+}
+
+// TestCancelTruncateTable tests cancel ddl job which type is truncate table.
+func TestCancelTruncateTable(t *testing.T) {
+	store, dom, clean := testkit.CreateMockStoreAndDomainWithSchemaLease(t, dbTestLease)
+	defer clean()
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t(c1 int, c2 int)")
+	defer tk.MustExec("drop table t;")
+	var checkErr error
+	hook := &ddl.TestDDLCallback{Do: dom}
+	hook.OnJobRunBeforeExported = func(job *model.Job) {
+		if job.Type == model.ActionTruncateTable && job.State == model.JobStateNone {
+			jobIDs := []int64{job.ID}
+			hookCtx := mock.NewContext()
+			hookCtx.Store = store
+			err := hookCtx.NewTxn(context.Background())
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+			txn, err := hookCtx.Txn(true)
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+			errs, err := admin.CancelJobs(txn, jobIDs)
+			if err != nil {
+				checkErr = errors.Trace(err)
+				return
+			}
+			if errs[0] != nil {
+				checkErr = errors.Trace(errs[0])
+				return
+			}
+			checkErr = txn.Commit(context.Background())
+		}
+	}
+	dom.DDL().SetHook(hook)
+	err := tk.ExecToErr("truncate table t")
+	require.NoError(t, checkErr)
+	require.EqualError(t, err, "[ddl:8214]Cancelled DDL job")
 }
